+++ src/main/java/org/monarchinitiative/lirical/cmd/YamlCommand.java
@@ -75,8 +75,8 @@ public class YamlCommand extends LiricalCommand {
                 .threshold(this.factory.getLrThreshold())
                 .mindiff(this.factory.getMinDifferentials());
         LiricalTemplate template = outputTSV ?
-                builder.buildGenoPhenoTsvTemplate() :
-                builder.buildGenoPhenoHtmlTemplate();
+                builder.buildPhenotypeTsvTemplate():
+                builder.buildPhenotypeHtmlTemplate();
         template.outputFile();
         logger.error("Done analysis of " + template.getOutPath());
     }
@@ -135,11 +135,7 @@ public class YamlCommand extends LiricalCommand {
         this.metadata.put("sample_name", factory.getSampleName());
         this.metadata.put("analysis_date", factory.getTodaysDate());
         this.metadata.put("yaml", this.yamlPath);
-        if (factory.global()) {
-            this.metadata.put("global_mode", "true");
-        } else {
-            this.metadata.put("global_mode", "false");
-        }
+
         Map<String,String> ontologyMetainfo = factory.hpoOntology().getMetaInfo();
         if (ontologyMetainfo.containsKey("data-version")) {
             this.metadata.put("hpoVersion",ontologyMetainfo.get("data-version"));
@@ -148,6 +144,11 @@ public class YamlCommand extends LiricalCommand {
         if (this.phenotypeOnly) {
             runPhenotypeOnly();
         } else {
+            if (factory.global()) { // global mode only makes sense for genomic analysis
+                this.metadata.put("global_mode", "true");
+            } else {
+                this.metadata.put("global_mode", "false");
+            }
             this.factory.qcExomiserFiles();
             factory.qcHumanPhenotypeOntologyFiles();
             factory.qcExternalFilesInDataDir();